refactored commonalites into basextractor
